fix: cycle imports and some other eslint warnings/errors
