+++ src/renderer/typings/styled-components.d.ts
@@ -3,6 +3,5 @@ import { darkTheme } from '../ui';
 type ThemeInterface = typeof darkTheme;
 
 declare module 'styled-components' {
-    // eslint-disable-next-line @typescript-eslint/no-empty-interface
     interface DefaultTheme extends ThemeInterface {}
 }
