+++ src/renderer/utils/auth.ts
@@ -1,0 +1,7 @@
+import { delToken, delUser } from 'renderer/stores/useAuthStore';
+
+export function signout() {
+    delToken();
+    delUser();
+    window.location.reload();
+}
