+++ src/renderer/utils/constants.ts
@@ -1,0 +1,2 @@
+export const isProd = process.env.NODE_ENV === 'production';
+export const isDev = process.env.NODE_ENV === 'development';
