+++ src/renderer/utils/rightNowInSecs.ts
@@ -1,0 +1,3 @@
+export function rightNowInSecs(): number {
+    return Math.floor(Date.now() / 1000);
+}
